V1/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.1
matplotlib>=3.4.3
tqdm>=4.62.2
PyYAML>=5.4.1